Auto reformat & code cleanup of all custom code

+++ Scripts/Nelderim/Complete Spell System/-=+ 01 Core/ScrollBag.cs
@@ -1,6 +1,8 @@
-using System;
-using Server;
+#region References
+
 using Server.Items;
+
+#endregion
 
 namespace Server.ACC.CSS
 {
@@ -10,25 +12,25 @@ namespace Server.ACC.CSS
 		{
 		}
 
-		public void PlaceItemIn( int x, int y, Item item )
+		public void PlaceItemIn(int x, int y, Item item)
 		{
-			AddItem( item );
-			item.Location = new Point3D( x, y, 0 );
+			AddItem(item);
+			item.Location = new Point3D(x, y, 0);
 		}
 
-		public ScrollBag( Serial serial ) : base( serial )
+		public ScrollBag(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
