+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ancient/Scrolls/MassMightScroll.cs
@@ -1,38 +1,36 @@
-using System;
-
 namespace Server.ACC.CSS.Systems.Ancient
 {
-    public class AncientMassMightScroll : CSpellScroll
-    {
-        [Constructable]
-        public AncientMassMightScroll()
-            : this(1)
-        {
-        }
+	public class AncientMassMightScroll : CSpellScroll
+	{
+		[Constructable]
+		public AncientMassMightScroll()
+			: this(1)
+		{
+		}
 
-        [Constructable]
-        public AncientMassMightScroll(int amount)
-            : base(typeof(AncientMassMightSpell), 0x1F62, amount)
-        {
-            Name = "Zwój masowej potęgi";
-            Hue = 1355;
-        }
+		[Constructable]
+		public AncientMassMightScroll(int amount)
+			: base(typeof(AncientMassMightSpell), 0x1F62, amount)
+		{
+			Name = "Zwój masowej potęgi";
+			Hue = 1355;
+		}
 
-        public AncientMassMightScroll(Serial serial)
-            : base(serial)
-        {
-        }
+		public AncientMassMightScroll(Serial serial)
+			: base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
-            writer.Write((int)0); // version
-        }
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
+			writer.Write(0); // version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
-            int version = reader.ReadInt();
-        }
-    }
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
+			int version = reader.ReadInt();
+		}
+	}
 }
