+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Bard/Scrolls/FoeRequiemScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Bard
 {
 	public class BardFoeRequiemScroll : CSpellScroll
 	{
 		[Constructable]
-		public BardFoeRequiemScroll() : this( 1 )
+		public BardFoeRequiemScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public BardFoeRequiemScroll( int amount ) : base( typeof( BardFoeRequiemSpell ), 0x14ED, amount )
+		public BardFoeRequiemScroll(int amount) : base(typeof(BardFoeRequiemSpell), 0x14ED, amount)
 		{
 			Name = "Soniczny Podmuch";
 			Hue = 0x96;
 		}
 
-		public BardFoeRequiemScroll( Serial serial ) : base( serial )
+		public BardFoeRequiemScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
