+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Bard/Scrolls/KnightsMinneScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Bard
 {
 	public class BardKnightsMinneScroll : CSpellScroll
 	{
 		[Constructable]
-		public BardKnightsMinneScroll() : this( 1 )
+		public BardKnightsMinneScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public BardKnightsMinneScroll( int amount ) : base( typeof( BardKnightsMinneSpell ), 0x14ED, amount )
+		public BardKnightsMinneScroll(int amount) : base(typeof(BardKnightsMinneSpell), 0x14ED, amount)
 		{
 			Name = "Wzmacniający Okrzyk";
 			Hue = 0x96;
 		}
 
-		public BardKnightsMinneScroll( Serial serial ) : base( serial )
+		public BardKnightsMinneScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
