+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Bard/Scrolls/SheepfoeMamboScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Bard
 {
 	public class BardSheepfoeMamboScroll : CSpellScroll
 	{
 		[Constructable]
-		public BardSheepfoeMamboScroll() : this( 1 )
+		public BardSheepfoeMamboScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public BardSheepfoeMamboScroll( int amount ) : base( typeof( BardSheepfoeMamboSpell ), 0x14ED, amount )
+		public BardSheepfoeMamboScroll(int amount) : base(typeof(BardSheepfoeMamboSpell), 0x14ED, amount)
 		{
 			Name = "Pasterska Przyśpiewka";
 			Hue = 0x96;
 		}
 
-		public BardSheepfoeMamboScroll( Serial serial ) : base( serial )
+		public BardSheepfoeMamboScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
