+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Druid/Scrolls/PackOfBeastScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Druid
 {
 	public class DruidPackOfBeastScroll : CSpellScroll
 	{
 		[Constructable]
-		public DruidPackOfBeastScroll() : this( 1 )
+		public DruidPackOfBeastScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public DruidPackOfBeastScroll( int amount ) : base( typeof( DruidPackOfBeastSpell ), 0xE39, amount )
+		public DruidPackOfBeastScroll(int amount) : base(typeof(DruidPackOfBeastSpell), 0xE39, amount)
 		{
 			Name = "Leśne Bestyje";
 			Hue = 0x58B;
 		}
 
-		public DruidPackOfBeastScroll( Serial serial ) : base( serial )
+		public DruidPackOfBeastScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
