+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ranger/Scrolls/HuntersAimScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Ranger
 {
 	public class RangerHuntersAimScroll : CSpellScroll
 	{
 		[Constructable]
-		public RangerHuntersAimScroll() : this( 1 )
+		public RangerHuntersAimScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public RangerHuntersAimScroll( int amount ) : base( typeof( RangerHuntersAimSpell ), 3828, amount )
+		public RangerHuntersAimScroll(int amount) : base(typeof(RangerHuntersAimSpell), 3828, amount)
 		{
 			Name = "Celność łowcy";
 			Hue = 2001;
 		}
 
-		public RangerHuntersAimScroll( Serial serial ) : base( serial )
+		public RangerHuntersAimScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
