+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ranger/Scrolls/SummonMountScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Ranger
 {
 	public class RangerSummonMountScroll : CSpellScroll
 	{
 		[Constructable]
-		public RangerSummonMountScroll() : this( 1 )
+		public RangerSummonMountScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public RangerSummonMountScroll( int amount ) : base( typeof( RangerSummonMountSpell ), 3828, amount )
+		public RangerSummonMountScroll(int amount) : base(typeof(RangerSummonMountSpell), 3828, amount)
 		{
 			Name = "Przyzwanie Wierzcha";
 			Hue = 2001;
 		}
 
-		public RangerSummonMountScroll( Serial serial ) : base( serial )
+		public RangerSummonMountScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
