+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Rogue/RogueBag.cs
@@ -1,7 +1,3 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Rogue
 {
 	public class RogueBag : ScrollBag
@@ -10,26 +6,26 @@ namespace Server.ACC.CSS.Systems.Rogue
 		public RogueBag()
 		{
 			Hue = 0x20;
-			PlaceItemIn( 30, 35, new RogueCharmScroll() );
-			PlaceItemIn( 50, 35, new RogueFalseCoinScroll() );
-			PlaceItemIn( 70, 35, new RogueIntimidationScroll() );
-			PlaceItemIn( 90, 35, new RogueShadowScroll() );
-			PlaceItemIn( 30, 55, new RogueSlyFoxScroll() );
+			PlaceItemIn(30, 35, new RogueCharmScroll());
+			PlaceItemIn(50, 35, new RogueFalseCoinScroll());
+			PlaceItemIn(70, 35, new RogueIntimidationScroll());
+			PlaceItemIn(90, 35, new RogueShadowScroll());
+			PlaceItemIn(30, 55, new RogueSlyFoxScroll());
 		}
 
-		public RogueBag( Serial serial ) : base( serial )
+		public RogueBag(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
